Remove the WritableBlock class and instances. (as in commit 4eaf251)
Stackingproof now has three admits:

Two of them are related to preservation of some parts of memory across external
calls.

The third is related to preservation of load across a store in the stack frame.
The argument used to be based on the WritableBlock predicates.

I'm not trying to prove them now as there are other commits by Tahina that I
need to incorporate first; and the proof of Stacking seems to change quite a lot
in the other commits.