Almost finished the whole simulation proof. Axioms left for external calls.
One axiom says that the permissions of the stack blocks are unchanged. This is
to be put as an additional property of the external calls.
The other states that the unique stack block's permissions are unchanged by
external calls. This is subtler because there is no obvious way of
characterizing this stack block except from naming it explicitly.