Simpler diagrams for forward simulations
We switch to a much simpler form for forward simulation diagrams, where
the callconv world is quantified existentially, but the related target
event is quantified universally. This made it possible to quickly verify
a number of additional passes, and reduce the number of changes needed
in order to do that.

This approach is made possible by embedding the queries and their
relatedness into the world itself. Thanks to a carefully chosen
interface, these details can largely be contained to the
LanguageInterface library.