No need for [eventval_valid] in [match_events]
It's only needed so that the definition of receptivity works with the
semantics of external functions, but we don't want or need receptivity
in the new approach.

As an added bonus, this means that [match_events], [match_events_query],
and [stable_event] don't need a global environment parameter any more.