Some progress on tailcallproof
I tried to change the extension into an injection because currently an extension
can only have a 'flat´ stack injection, which is not the case for Tailcall. The
problem is that I can't assume the injection is the flat injection because
that's not preserved by external call steps. (Actually it should but it's not
proven anywhere).

A possible step would be to generalize the memory injection function into an
arbitrary j but I would still need to know that the globals are in flat inj and
that the stack blocks inject into their target counterpart.

If that's too complicated, generalize the extensions in the memory model so that
they can be parameterized by a stack injection function, not necessarily the
flat stack injection.