Just one admit left in Inliningproof.
This is related to the case where a tailcall in the source program in inlined in
the target program.

Source:
int g(){
  // code of g
}
int f(){
  tailcall(g);
}

Target:
int f(){
  //code of g
}

With the accounting of stack-space usage, it is required that at every point of
the execution of the programs, the stack-space usage of the source program is
higher than or equal to the stack-space usage of the target program.

Technically, we need to prove that our oracle fn_stack_requirements is such
that:
fn_stack_requirements(g) >= fn_stack_requirements(f)

This is not however a constraint we can place on the compiler for the tailcall
pass because the oracle is generated by the compiler; hence the compiler cannot
depend on it.

I cannot at the moment think of an example where this situation would be
troublesome in practice, but I don't have (yet) a solution for the proof.